Make "default" optional if no activeNames
For those of us coming from a React Native background, having to specify a "default" property on every single style declaration is a little like nails on chalkboard. If the user is not specifying any activeNames, it's also completely redundant. Without it, the syntax is almost identical to React Native:

```javascript
import { StyleSheet } from 'react-native' // Basically the sames as `import reactcss from "reactcss"`

const styles = StyleSheet.create({ // Basically the same as `reactcss({...})`
  card: {
     backgroundColor: 'red'
  }
})
```

This is admittedly a syntactic nicety, but it completely backwards compatible and would just be an optional function signature.